requirements.txt
Django==4.2
djangorestframework
requests==2.32.3
pre-commit
types-requests
